Factor Selectors to separate method
